immediately parse stacks
